+++ src/sphinxsearch.cpp
@@ -7542,7 +7542,8 @@ ISphRanker * sphCreateRanker ( const XQQuery_t & tXQ, const CSphQuery * pQuery, 
 		{
 			// (word_docs > total_docs) case *is* occasionally possible
 			// because of dupes, or delayed purging in RT, etc
-			const int iTotalClamped = Max ( tSourceStats.m_iTotalDocuments, tWord.m_iDocs );
+			// FIXME? we don't expect over 4G docs per just 1 local index
+			const DWORD iTotalClamped = Max ( (DWORD)tSourceStats.m_iTotalDocuments, DWORD(tWord.m_iDocs) );
 
 			if ( !pQuery->m_bPlainIDF )
 			{
